feat: update DeletePedidoUseCase to return deletion status and validate existence of pedido

+++ src/applications/usecases/pedido/deletepedido.usecase.ts
@@ -6,7 +6,7 @@ export namespace DeletePedidoUseCase {
   export type Input = {
     id: number
   }
-  export type Output = void
+  export type Output = boolean
 
   export class UseCase implements DefaultUseCase<Input, Output> {
     constructor(private pedidoRepository: PedidoRepository) {}
@@ -15,11 +15,16 @@ export namespace DeletePedidoUseCase {
       if (!input.id) {
         throw new BadRequestError('Id é obrigatório.')
       }
+      const pedido = await this.pedidoRepository.findById(input.id)
+      if (!pedido) {
+        throw new BadRequestError('Pedido não encontrado.')
+      }
       try {
         const result = await this.pedidoRepository.delete(input.id)
         if (!result) {
           throw new BadRequestError('Pedido deletion failed')
         }
+        return result
       } catch (e) {
         throw new BadRequestError('Erro ao deletar pedido.')
       }
